make withSpeech hoc take props from parent so it doesn't need to know about state
