+++ client/src/components/hoc/withSpeech.js
@@ -1,24 +1,8 @@
 import React, { Component } from 'react'
-import { compose } from 'redux'
-import { connect } from 'react-redux'
-
-import { startListening, stopListening, addToRegister, clearRegister } from '../../actions/speech'
-import Mic from '../Mic'
-import '../styles/Speak.css'
 
 const head = arr => arr[0]
 
-const mapStateToProps = ({
-  speech: { isListening, register }
-}) => ({ isListening, register })
-
-const mapDispatchToProps = {
-  startListening,
-  stopListening,
-  addToRegister,
-  clearRegister,
-}
-
+// withSpeech :: Required Props: startListening, stopListening, addToRegister (all functions)
 const withSpeech = Comp => {
   return class WithSpeech extends Component {
     componentDidMount = () => {
@@ -62,37 +46,5 @@ const withSpeech = Comp => {
   }
 }
 
-const CreateNoteWithSpeech = ({
-  isListening,
-  register,
-  start,
-  stop,
-  onSave
-}) => {
-  return (
-    <div className="Speak">
-      <Mic
-        startListening={start}
-        stopListening={stop}
-      />
-      {isListening
-        ? null : (
-          <button
-            onClick={() => onSave(register)}
-            className="saveSpeechBtn mainBtn"
-          >
-            Looks good?
-          </button>
-        )
-      }
-      <div className="transcript">
-        {register.map(t => <p key={t}>{t}</p>)}
-      </div>
-    </div>
-  )
-}
 
-export default compose(
-  connect(mapStateToProps, mapDispatchToProps),
-  withSpeech,
-)(CreateNoteWithSpeech)
+export default withSpeech
